Fixed Bug #{1845}
The mapref(n) "box" and "dim" methods were erroneously using the "c2m"
method to compute the bounds in map coordinates.  However, this method
only works when the zoomFactor is 100%.  The bounds in map coordinates
are now computed when the map dimensions are set, and simply reported by
"box" and "dim".