Fixed Bug #{1859}
After due reflection, I've decided that the GUI update issue is not a problem.

Consequently, this package updates nbhood(sim) and civgroup(sim) to delete
their dependent neighborhood groups when a neighborhood or civgroup is
deleted, and also updates a host of man pages to cover the additions of 
the past week.