Fixed Bug #{1878}
Made the following changes:

On "order send", the order handler is wrapped in a transaction unless
the order is sent by a test script (in that case, it might be useful
to examine the database to see what's going wrong).

When "order send" is sent by the GUI, and there's an unexpected error
in the order handler, the error is logged immediately, and a detailed
error message is displayed to the user.  It says:

    There was an unexpected error during the 
    handling of this order.  The scenario has 
    been rolled back to its previous state, so 
    the application data  should not be 
    corrupted.  However:

    * You should probably save the scenario under
      a new name, just in case.

    * The error has been logged in detail.  Please
      contact JPL to get the problem fixed.

Then, the order processor calls "scenario reconfigure"; this causes
the entire app to reconfigure itself from the database, cleaning up
any problems in Tcl memory.

A similar process occurs when an error is encountered while 
undoing an order.