Fixed Bug #{1884}
Added a *:UPDATE:MULTI order for each *:UPDATE order, and added the
relevant test cases to the order test suites.

Added a "-listof" option to "prepare"; "-listof nbhood" says that the
value is a list of nbhood values.  Added appropriate "validate"
methods to the nbgroup and sat modules, which had previously lacked
them.

Removed the "-trim" option from "prepare", as it used in every single
case.  Updated all of the calls to "prepare" accordingly.