Fixed Bug #{1884}
Multi-update has been implemented in all of the entity browsers.  If you
select one or more items, the pencil button is enabled.  When it is pressed,
the order dialog varies depending on how many items are selected.  For 
one item, the normal update dialog appears; for multiple items, the 
multi-update dialog appears.  

The multi-update dialog indicates how many items have been selected; it
does not list them (as there could be hundreds).