Fixed Bug #{1908}
Updated the order definition syntax slightly; the "table" command is replaced
by the "options" command, so that there can easily be an expanding set of
order options, including "-table".  Added "-tags", so that each order can
have tags associated with it; this allows clicking on a row in a browser to
completely update the related order dialogs.