Fixed Bug #{1923}
Added a new mapicon(i) type, mapicon::unit, which supports both the
friend, enemy and neutral shapes, but also a variety of symbols.  The
mapviewer now draws the unit icons with the appropriate shape and symbol
for the existing units.

Much remains to be done; the mapviewer currently only updates unit icons
on refresh, for example.