Fixed Bug #{1846}
This turns out to be a simple GUI update error.  There's a
variable that contains the zoom factor for display; however,
on refreshing the mapviewer, as happens when a new scenario is 
opened, the wrong variable was updated.

This is now fixed.

It's an open question whether the display should jump back to 
100% zoom when the display is refreshed; however, I tried making
it retain the current zoom factor, and when it's larger than 100%
this caused an ugly delay.  I've therefore left the behavior as it
is.