Fixed Bug #{1958}
Moved the "scenario reconfigure" command to sim(sim) as "sim reconfigure".
The scenario object was the most central object to the simulation...until
the sim object was created.  However, its real job is simply managing the
"new", "open", and "save" operations.  Reconfiguring the app and GUI is
really a responsibility of the simulation.

As part of this, of course, the <Reconfigure> event is now published by
::sim, not by ::scenario, so all subscribers needed to be updated.