Fixed Bug #{1962}
Each order can now be associated with particular application states,
and order(sim) is apprised of the current application state.  Orders
are rejected if sent in the invalid application state.  Plus,
order(sim) sends a <State> event when the state changes; the future GUI
changes can use this to enable/disable controls when the state changes.

The order(sim) man page was updated accordingly.