Fixed Bug #{1963}
Done.

Defined statecontroller(n), with test suite and man page.
Defined the four new cond::* statecontrollers.
Updated all widgets that cause an order to be sent or popup its
dialog to be controlled by one of the controllers.
Updated the File/New Scenario and File/OpenScenario to be controlled
as well.