Fixed Bug #{1969}
I've changed the name "WAYBACK state" to "SNAPSHOT state" because, properly
speaking, the "Wayback Machine" allows you to go back and *change* the past,
and the "SNAPSHOT" state doesn't.  Thus, "SNAPSHOT" is more descriptive.

I've also improved the various messages that appear in the GUI, and added
a tick counter to the simulation toolbar.

Finally, I've checked the existing orders; only MAP:IMPORT is useful during
the SNAPSHOT state, and the UNIT:* orders, which previously were always valid,
are now valid in every state *but* the SNAPSHOT state.