Fixed Bug #{2040}
Added the activity and activity_gtype tables to scenariodb.sql, and 
populated them in scenario(sim).

Added the activity(sim) module, to handle validation and queries, and
the computation of activity coverage.

Added most of the parms required for computing activity coverage.

Added most of the code to compute coverage.