Fixed Bug #{2040}
Added a new parmdb(5) parameter, activity.*.assignToActive, which is the
ratio between the number of people assigned to an activity, and the
number who will actively be performing it at any given time, given
the implied schedule for the activity.  It defaults to 1.  Updated
the application to make use of it.