Fixed Bug #{2040}
Done.  Added browsers and browser entries for the activity and security
data.  While so doing I fixed a number of problems.  The sim(sim) code
was initializing GRAM et all *before* creating the tick 0 snapshot, which
is the opposite of what I intended; and then, the snapshot mechanism was
ignoring empty tables.  Thus, force_ng was empty in the tick 0 snapshot...
and so returning to the tick 0 snapshot left force_ng containing whatever
it had held before returning to the snapshot.  This was very confusing.