Fixed Bug #{2098}
Fixed problem: when a neighborhood is raised or lowered in the stacking
order, the mapviewer(n) was drawing a new set of neighborhood polygons...
without deleting the old neighborhood polygons.  This left two polygons
on the map for each neighborhood; and when the neighborhood was moved,
the old polygon stayed in place.