Fixed Bug #{2164}
This implements a preliminary version of the aam(n) module and, in
particular, the "aam mutate attrit" command, which will be used by 
the ATTRIT:GROUP order.

However, after doing it I realized that the algorithm for civilian
attrition isn't quite right.  I've worked out the details, and I
need to update the draft Analyst's Guide before moving ahead.  So
I'm checking this in to save it.