Fixed Bug #{2194}
Done.

The ATTRIT:* orders each take two new arguments, g1 and g2, for one or
two responsible force groups.

Instances of attrition to org groups are recorded as entries in attrit_nf.

Instances of attrition to civ groups are recorded as entries in attrit_nf.

If the CIV attrition was due to one or more force groups, entries are
also made in the attrit_nfg table.