Fixed Bug #{2197}
Done.

The attacking ROEs for Uniformed Forces will cause attrition to the
defending Non-uniformed forces, with collateral damage to the civilians
in the neighborhood.

Note that you can see what AAM is doing by going to the log, filtering
on AAM, and selecting the "detail" log level.  This will show you not
only what attrition is taking place, but why an ATTACK ROE might be
having no effect.