Fixed Bug #{2199}
Added a "delete" command to reporter(n), so that undone reports can be
deleted.

Added a "-deletecmd" callback to report(n), so that the GUI can be updated
when a report is deleted.

Added an "Attitudes" bin with a "Drivers" sub-bin to the report browsers.

Added the "Attitude Drivers" report with a new REPORT:DRIVER order to
request it.

Added a new menu, Reports, to the main window.

Added a new section, "Reports", to the on-line Help; it can be reached
by selecting Help/Reports Reference.  In addition, for user-requestable
orders the "Help" button on the report's order dialog will bring up the
report's page, as usual.

Added a help page for the Attitude Drivers Report.

Added some helpful macros to help(5).