Fixed Bug #{2218}
It is now possible to set and update the auto-resolution duration
while an ensit is in the INITIAL state.  Further, the ensit module
scheduled auto-resolution.

Added tests to test auto-resolution, as well as the clean-up and
restoration of the auto-resolve event on manual resolution and undo.

Finally, the ensitbrowser(sim) module shows the resolution time.