Fixed Bug #{2222}
Added the ORDER:SCHEDULE and ORDER:CANCEL commands, which schedule
orders and cancel scheduled orders respectively.

Added the -canschedule order option to order(sim), and defined it for
each order that can be scheduled.

Added the new orderbrowser(sim) browser, which shows the scheduled
orders; it appears on the new Orders tab.  The user can browser the
scheduled orders, and cancel them.

There is at yet no GUI for scheduling orders; in addition, the
tests have not yet been written.