Fixed Bug #{2287}
athena_sim(1) calls a routine that's defined in bin/athena.tcl (the
loader script) to log loaded mods.  This routine is not available
when athena_test(1) is executed, and so "make test" failed.  For now,
I updated athena_sim(1) to call the routine only if it is defined;
athena_test(1) doesn't support mods.