Fixed Bug #{2290}
Done.

The behavior of Tkhtml2 is odd.  If no -isvisitedcommand is specified,
then the -visitedcolor and -unvisitedcolor options are ignored and all 
links are blue.  If the -isvisitedcommand is specified, then the values
of these options are used, and all links are black because that's the
default value for both of them.

This commit does two things:

1. htmlviewer(n) sets the default -visitedcolor and -unvisitedcolor
   values to be consistent with the Firefox defaults.

2. helpbrowser(n) sets the -isvisitedcommand, and remembers whether
   URIs have been visited or not.