Fixed Bug #{2424}
Created a skeleton for econ(sim).  It is:

* Initialized by app(sim)
* Calibrated by sim(sim) on leaving the PREP state.
* Advanced by sim(sim) at each econ.ticksPerTock.

In addition, I found a bug in appwin(sim):  it was assuming that 
SIM:RUN was doing the scenario sanity check, when it's SIM:LOCK that does 
it.  Consequently, if SIM:LOCK's sanity check failed then Athena got a 
bgerror.  This is now fixed.