Fixed Bug #{2452}
Fixed three errors in the view(sim) code:

1. The neighborhood variables with no arguments (e.g., "nbmood") were
throwing errors because there was no validation command.

2. "view n validate" returns the canonicalized variable name; but for
names like "nbmood" it was adding an extraneous ".".

3. The validation code for cov.<g>.<a> was using the wrong set of 
activities.  "NONE" is assignable for all three group types, but has
no coverage; "PRESENCE" is not assignable for any group type, but has
coverage for force groups.  Fixed this by adding a number of types to
ptype(sim).