Fixed Bug #{2467}
Done.

To test, create a scenario, and push the "Prep Lock" button.  Then,
enter this command:

   nbfill cov.<g>.presence

where <g> is a force group name.  Assuming that there are no force groups,
all neighborhoods should be white.  Then, create a unit belonging to 
group g.  Move it to a neighborhood, and press the new "A" button on the
simulation toolbar.  The neighborhood the group is in should be shaded
blue, with the darkness correlating with the coverage.

Try moving the unit from neighborhood to neighborhood, and changing its
personnel, pressing the A button in between each time.