Fixed Bug #{2470}
This bug results from the change in the order of events in a time tick
(bug 2398).  

The gram_contribs (history of contribution by driver) table
is not included in the snapshots; instead, when "changing the future"
at a snapshot, all gram_contribs from the future are deleted.  The
test is "time > now"; but with the new tick scheduling, it needs to
be "time >= now".