Fixed Bug #{2543}
Fixed.

Of all the stupid things, "demog analyze pop" was being called only
at simulation start, and when returning to the PAUSED state after
advancing time.  Thus, if you ran for 15 days at once, and the
demographics changed, the results weren't seen until the *next* run.

"demog analyze pop" is now being called at the beginning of each
tick's analysis, as it ought to be.