Fixed Bug #{2626}
Implemented the send command as described.  As a bonus, parameters
with default values (specified as -defval in the parameter meta-data)
can be omitted from the send.  When an order is rejected, the error message
includes both the specific errors and a list of parameters that were 
defaulted.  Thus, typing (say) "send FRCGROUP:CREATE" will yield most
of the information you need to know.