Fixed Bug #{2630}
Found and fixed an error in computation of the LSF in econ(sim) while
testing this.  The error resulted from the removal of the nbgroup tables,
and is not present in Athena 2.1.

Added the export command, which saves the order history as a script given
a file name.  The file is saved relative to the current working directory.
The SIM:* orders need to be handled specially.

* SIM:UNLOCK appears in a vacuum: unlocking the sim chops off the 
  order history just before the SIM:LOCK, and then saves the SIM:UNLOCK.
  So SIM:UNLOCK needs to ignored.

* SIM:RUN should always have "-block yes", so that the next command in the
  script won't execute until the simulation pauses.

* SIM:PAUSE means that the previous SIM:RUN was terminated explicitly by
  the user.  So instead of writing SIM:PAUSE to the script, update the
  previous SIM:RUN to stop at the correct time.