Fixed Bug #{2306}
Updated Athena to use the new "colorfield" data entry field.  This
field provides a palette of standard colors for the user to pick from,
all of which are designed to work well as unit colors.

The color picker also allows the user to pick any color they like,
so it doesn't eliminate the problem described in this bug; but it makes
it much easier for them to do something sensible.