Fixed Bug #{2660}
The problem is that view(sim) didn't flush its view cache when the
current scenario is discarded and a new blank one is created.
view(sim) was designed to flush its cache when the scenario is restored.
HOWEVER, there is no restore when the current scenario is thrown away
and a new one is created, thus leaving a loophole for this bug to
escape through.

As of now, view(sim) no longer registers as a saveable(i) object;
instead, it flushes its cache on the ::sim <DbSyncA> event.