Fixed Bug #{2751}
Completed the GUI portion of the integration.  The Groups/Relationships
browser now shows all computed relationships in green; relationships 
overridden by entries in the rel_fg table are shown in red.  The 
REL:UPDATE:* orders have been replaced by REL:OVERRIDE:* and 
REL:RESTORE; the former creates or updates records in rel_fg, and the
latter deletes records in rel_fg.  The browser has buttons for
both of these.  In addition, the browser now has a view-selection
pulldown; you can see all relationships or only the overridden ones.

Since rel.gg is always 1.0 and cannot be changed, it is no longer shown
in the browser.