Fixed Bug #{2869}
Done.  Most of the updates were to the appwin(sim) module.  Also, I removed 
the Plan/Force Levels browser, the EnSit browser, and the MAD browser, as the
related capabilities are going to be removed or replaced.  I rearranged
the tabs to some extent.  Finally, I updated the Help pages in a minimal way; 
they are going to need a thorough scrub once we're closer to being finished
with Athena 3.