Fixed Bug #{2895}
I've removed the code that remembered which tab you were on, as I find it
annoying rather than helpful.  It would be nice to save the mode and tab
when saving the scenario to disk, but the same code is used when taking
snapshots, and having the tab saved in the snapshot is too annoying for
words.

At some later point, I might try to revamp the checkpointing logic so that
some objects are checkpointed only when saving the scenario to disk.