Fixed Bug #{2909}
Added fields to vrel_ga and bvrel_tga so that we can give an analysis
of the current vertical relationships.

Updated control(sim) to populate these fields.

Updated appserver(sim) with the new /group/{g}/vrel URL and handlers,
and put a link to it on the /group/{g} page for civilian groups.

(Force and ORG groups don't have summary pages yet.)