Fixed Bug #{2913}
Implemented as described.  In addition, the myserver(n) interface for
content handler routines has been changed slightly.  Previously, 
content handlers got two arguments: the URL given to the server, and the
name of an array of pattern matches.  Now, the URL argument has been
replaced by an URL dictionary, or "udict".  This dictionary contains the
result of passing the URL to Tcllib's uri::split, with one additional key:
the value of "url" is the original URL.  This allows handlers to get
the query portion of the URL without reparsing it.