Fixed Bug #{2985}
The problem was absurdly simple, and had nothing to do with the 
MET/UNMET conditions.

When the "mode" is "ALL", the DEPLOY tactic computes the maximum
of troops the actor can afford to deploy: 

   max_troops = cash_on_hand/cost_per_troop.

The actual number of troops deployed is then

   troops = min(available, max_troops)

However, troops is then a floating point number, and we want an integer; so
we compute 

   max_troops = int(cash_on_hand/cost_per_troop)

Unfortunately, the int() function has a deplorable property: it
converts the input to a 32-bit number (64-bit for 64-bit builds).  If
cash_on_hand is large enough, which it can be, this yields an
integer overflow, and a negative result.  We then deployed a negative
number of troops, which corrupted the data.  The actual bgerror occurred
the next time around.

This should *not* be a problem, given that Tcl supports "bignums": 
arbitrary precision integers.

The answer seems to be this:

   max_troops = round(floor(cash_on_hand/cost_per_troop))

floor(x) rounds the number down to the nearest integer, while leaving
the result in floating point form.  Then, round(x) converts to the 
nearest integer.  Together, they do the job that int(x) ought to do.

I also added an assertion to "personnel deploy" so that it complains
if given a non-positive number of troops; this will catch similar errors
earlier.