Fixed Bug #{2987}
The condition was storing the number of troops in a floating point
field rather than an integer field.  I've added an "int1" column to the
conditions table; the TROOPS condition now uses this instead of "x1".
As a result, the number will appear as an integer in the
CONDITION:TROOPS:UPDATE order.