Fixed Bug #{2995}
Both the CASH and INFLUENCE conditions had the same problem: the values
in question are rounded for display but were not rounded prior to
comparison.  Thus, values that look like they should be equal might not
be.

The fix in both cases was to round the numbers before comparison, to 
two decimal places in both cases.

The help pages for both conditions were updated as well.