Fixed Bug #{3037}
Fixed.  The underlying parmset(n) package doesn't have a "validate" 
method, so the PARM:SET order was simply trying to save the parameter
value by calling "parm set", and rejecting the order if "parm set"
threw an error.  This worked fine when PARM:SET was used only to 
implement an executive command, but the order dialog has stricter
needs.

The code now retrieves the parameter's validation type, and uses it
explicitly to validate the parameter value.