Fixed Bug #{3038}
As part of this, I recast the "Reports" tab as the "Firings" tab;
the only reports that will appear on that tab are rule firings.  Then,
I made the Firings tab visible only on request; it's hidden by default.
Finally, I deleted the irrelevant bins from the report browser in the
Firings tab, and the "hotlist" menu items from the Reports menu.

In addition:

* REPORT:* orders are not written to the CIF.
* The Report/Model Parameters report appears in the detail browser.
* I updated the on-line help accordingly.