Fixed Bug #{3053}
Fixed.  The problem is that temporary tables were being included in
scenario snapshots by scenariodb(n)'s tclexport routine.  The
temp_satcontribs table, in particular, is sometimes there and sometimes
not there; and in this case, it was there when the snapshot was taken
but not when the snapshot was loaded again, causing the bgerror.

Practically speaking, temporary tables shouldn't be included in the
snapshot to begin with.  Temporary tables are either constant look-up
tables or transient tables used only for the duration of some particular
algorithm; either they don't change, or they don't matter.  So the fix
is to update scenariodb(n)'s tclexport code to include only permanent
tables.