Fixed Bug #{3039}
Done.  The Reports/Attitude Drivers... menu item now displays the 
desired report in the Detail browser.

In addition, there was a bug in the report code: start and end times
were not being displayed for "inactive" drivers.  This has been 
fixed; the start and end times now reflect the actual contributions
made by the driver.