Fixed Bug #{3064}
Fixed.  The logic in the TACTIC:ATTROE:UPDATE order handler was wrong:

1. Only the tactic_id parameter is required.
2. The optional parameters may be omitted.
3. All parameters that are not missing are validated at the top of the
   handler.
4. The handler needs all of the tactic's parameters to finish
   validating the update, so it retrieves any that are missing.
5. The handler assumes that all of the retrieved values are already valid.
6. For tactics and conditions, this isn't a valid assumption.

The fix was to validate the tactic_id completely, then retrieve the
missing tactic parameters, and then do all of the rest of the validation.
This guarantees that an invalid or missing parameter is detected before
it can cause a problem.