Fixed Bug #{3083}
Updated the mapviewer(n).  Whenever a unit record is updated, mapviewer(n)
calls UnitDraw to redraw the unit.  UnitDraw was ignoring the unit's 
"active" flag; consequently it was drawing both active and inactive units.
The solution was to make UnitDraw simply delete icons for units that are
inactive, but draw active units normally.