Fixed Bug #{3115}
Just after committing the previous set of changes, I realized that the 
ENI rule set, as designed, wasn't going to work: if the group was 
receiving saturation, then there would be a 2.0*M+ effect on every concern
indefinitely.  That's not good.  After discussion with RGC, we determined
that the "needs" term only applies in the ENI rule set when needs < 0.0.

I handled this by breaking the single rule into four rules: 

* ENI is less than required
* ENI is less than expected
* ENI is as expected
* ENI is more than expected

The latter three all have the same effect, but they document the case
that obtained as well.

Updated the rules document again, and updated the test suite.