Fixed Bug #{3129}
Done.

To test, create EXECUTIVE tactics for an actor and for the SYSTEM agent.
The tactic can use the "sigevent" command to write to the Significant Events 
log.  Test the "Once" parameter, per the on-line help for this tactic.

Also, test error handling using the "call" command.  Create a script in
your working directory, "test3129.tcl":

    log "Testing Bug 3129"
    # error "Simulated Error"

The EXECUTIVE tactic can call this script using the command

    call test3129.tcl

It should run without error, writing "Testing Bug 3129" to the debugging log.
Also, look at the Significant Events log.

Then, edit the script, removing the "#" on the second line, and let the
tactic execute again.  The Sig Events log should say that the tactic
failed due to a "Simulated Error"; the debugging log should say the same,
and should include a stack trace.