Fixed Bug #{3131}
Updated tactic_executive(sim); it now sets the order state to TACTIC
before calling the executive command, and returns it to its old value
after.

Added TACTIC to the -sendstates for CIVGROUP:UPDATE:POSTPREP and to
ATTRIT:NBHOOD.  The latter has issues, which will be dealt with when I
enable the other attrition orders.

Added the "tactic" order interface in app(sim).

Updated the "send" command to use the "tactic" interface when appropriate.