Fixed Bug #{3153}
Done.  There were many changes; here are the most significant ones:

1) Each actor now has a "supports" attribute, which names the actor to
whom they usually give political support across the playbox.  The value
is an actor's name, "SELF", or "NONE".

2) The actor can modify this default setting neighborhood-by-neighborhood
using the SUPPORT tactic.

3) Each actor now has "direct support" in each neighborhood, computed just 
as support used to be, and also "total" or "actual" support, which is the 
total of the support values from the actors that support this actor in this
neighborhood.  By default, this is just the actor's own direct support.
Thus, an actor can add to another actor's influence, possibly allowing that
actor to gain control of the neighborhood.

4) The FUNDENI tactic now requires that an actor have a minimum direct
support in the neighborhood, rather than positive influence--that way, an
actor who lends his support to another actor can still fund ENI.