Fixed Bug #{3123}
When a new agent was selected in the Agent list, the code did a complete
reload of the browser's content--which reloaded the Agent list.  But the
agent selection code should only be triggered by a complete reload or by the
user selecting an agent; in neither case is it appropriate to reload everything.
Instead, the routine now just reloads the goals and tactics.