Fixed Bug #{3191}
Done.  Added an "Auto-Calc" checkbox and a "Calc Now" button to the
Belief System browser's main toolbar.  Affinities will be recalculated
on any change so long as the checkbox is checked.

The only differences from the original description are the name of the
"Calc Now" button and that the "Calc Now" button is not disabled when the
Auto-Calc box is checked.