Fixed Bug #{3265}
Done.  To test, browse the Athena docs tree, and verify that the Mars
Analyst's Guide is listed under the Athena Analyst's Guide, and that
clicking on the link opens it.  Note that you'll need to have OpenOffice
(or, these days, LibreOffice is probably a better choice) to be able to
open it.