Fixed Bug #{3316}
The Econ/Overview tab uses the cmsheet(n) widget, which is based 
on the TkTable widget.  This widget does not resize columns to fit their
content, and the fonts are just enough different on windows that some
columns were two narrow.  This change expands the relevant columns
generously, so that everything will appear.