Fixed Bug #{3281}
Fixed.  The problem is in the htmlviewer3(n) module, and occurs when you
do this:

1. Highlight some text in the Detail Browser.
2. Let time advance (or do something else that causes the browser contents
to refresh).

The interaction data used when sweeping out a selection is not cleared,
but the "nodes" it refers to no longer exist.  The solution is to clear 
the interaction data while clearing the browser as part of the refresh.