Fixed Bug #{3235}
Added cond::simNotRunning to the File/Save Scenario and 
File/Save Scenario As menu items (File/Open Scenario and File/New Scenario
already had this).  This ensures that the menu items are disabled when
RUNNING.

However, there are also hot-keys.  Consequently, I added checks for 
{[sim state] eq "RUNNING"} to appwin(sim)'s FileNew, 
FileOpen, FileSave, and FileSaveAs routines, so that they are no-ops if 
invoked by hot-key while RUNNING.

Then, there are the "load", "save", and "new" executive commands.  These
errored out with an assertion failure if the state was RUNNING; I changed
the [assert] to a [require] so that it would give a nicer error message.