Fixed Bug #{3157}
The econ model is now calibrated and initialized just before the first
strategy tock, rather than after.  This means that the economic results 
displayed in the Econ/Overview tab on lock will reflect the status quo, 
rather than the on-lock strategy execution.  But since we usually update
the economic model just prior to strategy execution, this will usually
be the case at any time.