Fixed Bug #{3358}
Fixed!

The ruleset is changed and the specifics can be found in the updated rules document. In short, CUL and SFT are no longer effected and AUT and QOL are effected less.

Also, the needs factor now runs from 0.0 to 2.0 (nominally) using these conditions:

* needs is 0 when actual level of service (ALOS) >= required level of service (RLOS)
* needs is 2.0 when ALOS = 0.0
* needs varies linearly between the two.

Suggested tests:
Using the FUNDENI tactic set up a scenario where all four rules (ENI-1.1 to ENI-1.4) fire.
Verify that CUL and SFT are no longer effected and that AUT and QOL effects are in line with the new magnitudes.
Verify that given an ALOS and RLOS that the needs factor is computed as described.