Fixed Bug #{3365}
Done!

Status quo funding is removed from Athena. The saturation and required levels of funding now appear in the detail browser both under the Civ. Groups object and in each individual civ group (once Athena is locked).

To test:
Inspect the detail browser for these levels of funding both in the PREP and LOCKED states. Have the population change (magic attrition works well) and let Athena run to the next tock. Verify that the levels of funding reflect the new population.