Fixed Bug #{3371}
Done!

All create/update tactic order dialogs and accompanying helps have been updated to include this new flag.
The strategy browser now shows the on lock flag as well.

The strategy logic is updated to look at the on lock flag and if it is "YES" and the scenario is locking, the tactic is deemed eligible for execution and an attemps is made to execute it. "On lock" tactics are still subject to resource constraints.

Suggested tests:
* Create some tactics with the "Exec On Lock" flag set to "YES".
* Lock the scenario and verify by looking in the logs that an attempt was made to execute any and all "on lock" tactics.
* Inspect the on-line helps to verify that the "Exec On Lock" field is properly documented.