Fixed Bug #{3381}
This is a minor change to where the simclock tick size gets set. Rather than in the sim object, it's at the app level. It's not really something that the sim should be able to change/control since it's not something that will be configured for anything other than 1 week.

The previously suggested tests still apply.