Fixed Bug #{3423}
Athena is now set up to do a complete "pseudo-tick" when locking the scenario.
It initializes all modules, executes the on-lock strategies, and does all of 
normal analyses and assessments, except for those that produce persistent
effects, or which would have no effect at t=0 anyway.  Thus, no attrition
occurs during lock, neighborhood control does not shift, and the MOOD rule
set is not called because mood hasn't had time to change.

In addition, the service(sim) module is fixed; the Expected Level of Service
(ELOS) needs to be set to the Actual Level of Service (ALOS) on lock, and
because the initial ALOS is being determined at a different time in a different
way (i.e., by on-lock FUNDENI tactics) the ELOS wasn't being set properly.

Updated tests as appropriate.