Fixed Bug #{3429}
Done.  There were two issues here.

First, the ensit.tcl module was calling both the monitor rule subset 
(e.g., DISASTER-1-*) and the resolution rule subset (e.g., DISASTER-2-*)
at the tick when the ensit was resolved.  This was correct Athena 3, 
since the monitor subset needed to terminate the ensit's on-going
slope effects.  It has been changed, so that when the ensit is resolved,
only the resolution rule set is called.

The second issue that resolution only has an effect for some ensit types,
and then only when there is a resolving group that is local to the playbox.
Otherwise the rule set will be triggered, but no rules will fire.