Fixed Bug #{3430}
Done.  Made ensit(sim) a little smarter about assigning the start time for
ensits.  If in PREP, the start time is 0; if PAUSED at time t, the start time
is t+1; otherwise, we're RUNNING, and whether the ensit is created by a tactic
or by a scheduled order the start time is t.  In every case, this is the 
time of the first assessment.