Fixed Bug #{3439}
Done; this was a straightforward change to engine(sim)'s LoadAram routine.
HREL and VREL curves are passed the natural level, as provided by the
hrel_view and vrel_view.  SAT curves are passed 0.0 as the natural
level at present, because they mostly don't have one; any value other than
0.0 is misleading.  COOP curves are passed the baseline value as the natural
level at present, because this is consistent with JNEM.