Fixed Bug #{3428}
When the sanity check fails for a tactic, as it does in this case, 
the tactic is marked "invalid".  The strategy browser was attempting to 
get the estimated cost of the invalid tactic; and the tactic was trying 
to retrieve the activity cost for the group type, but since there was
no group, there was no group type, and the attempt failed.

The error here is that the strategy browser was attempting to determine the
cost of an invalid tactic.  This has been fixed; in such cases, the cost
will display as "???".