Fixed Bug #{3413}
Fixed.

The browser page with the error is obsolete; it's trying to show all 
of the drivers that affect vertical relationships in Athena 3,
and that has changed considerably.  Thus, I deleted the page instead of
fixing it.  However, I ran into an additional problem, where the civgroup
page was referring to gui_vrel_view when it should be using gui_uram_vrel. 
I fixed these.