Fixed Bug #{3447}
Done!

On lock is no longer an option for these tactics:
ATTROE, DEFROE, DEMOB, FUND, MOBILIZE, SAVE and SPEND.

On lock remains an option for these, and for the ones that did not have a "Once" option it's been added:
ASSIGN, DEPLOY, DISPLACE, EXECUTIVE, FUNDENI and SUPPORT.

The automated test suite is updated and all the online documentation has been updated as well.

Suggested tests:
* Create at least one of each of the tactics and verify that the "On Lock" and "Once" options are in place (or not in place) as described.
* Edit the created tactics and verify that "On Lock" and "Once" can be toggled after creation.
* Inspect the documentation for completeness and correctness.