Fixed Bug #{3455}
Done.

Unowned CAPs now cause a sanity check failure. 
While testing this, ran across a different problem: CAP owners were not being
restored to the caps table if their deletion was undone.

This has also been fixed and a couple of tests added to the test suite to test this case.

To test:
Create a CAP with an owner that is subsequently deleted. Lock the scenario. 
Inspect the detail browser for an appropriate message and directions for how
to fix the failure.