Fixed Bug #{3498}
The core is implemented. Semantic hooks and associating topics with semantic
hooks is now in.

The automated test suite has been updated with tests for the new functionality.

The GUIs still need to be implemented and semantic hooks still need to be
associated with payloads as part of an information operations campaign.