Fixed Bug #{3518}
Done.

It takes a hook ID as an argument and returns a dictionary of topic/position pairs.

If the hook is not defined or if none of the topics are "normal", the empty string is returned.

This will be used by the IOM module and will get exercised there. No specific I&T under this bug.