Fixed Bug #{3523}
Done!

Now, when the scenario is locked the list of actors that have access to a CAP is listed on both the CAPs list page and the individual CAP page. The actors that have access appear as hyperlinks as described: owner first, everyone else afterwards in alphabetical order.

To test:
Grant access of a CAP to other actors using the GRANT tactic.
Lock the the scenario and view the all CAP pages. Verify that the access list is visible and appears as described.