Fixed Bug #{3529}
Done!

In the individual detail browser pages for semantic hooks there is now a section on which IOMs use that hook.

Tests:
Create at least one semantic hook and some IOMs that use it.
Verify that on the detail browser page for the semantic hook the IOMs that use it are listed properly.