Fixed Bug #{3496}
Done!

The CAP browser now has three panes laid out like this:
  * The CAPs themselves in a pane at the top
  * The Neighborhood coverage and CAP coverage browsers below the caps side by side
  * The neighborhood coverage browser is to the left of the CAP coverage browser.

When one or more CAPs are chosen in the top pane only the data pertaining to those CAPs are displayed in the lower panes.
If no CAPs are selected then all the data is shown in the lower panes. This means that the lower panes automatically filter on the selected CAP(s).
There are no filter boxes in the lower panes because of this.

To test:
* Create a number of CAPs and verify that the behavior is as describe above.
* Update and delete CAPs, coverages, and penetration.
* Verify that the data is consistent across all panes.