Fixed Bug #{3484}
This behavior is encountered only if Athena has not been saved. 

What it's doing is trying to popup a message box prompting the user to save or discard the scenario.
The problem is the message box has no parent to popup over because it is iconified.

The fix is to deiconify the window if it's not been saved.

To test:
* Iconify Athena in an unsaved state and try to close it.
* Verify that Athena deiconifies and the message box appears.
* Iconify Athena in a saved state and try to close it.
* Verify that it closes without problem.