Fixed Bug #{3631}
This particular problem is fixed. That is, to say, the one that the attached scenariodb file causes.

There is, however, a wider problem that will be addressed. The new economics model will need to have more sanity checks done on it as time goes on.

This commit introduces the first of those. Namely, that A.goods.pop (aka the per capita annual demand for goods) must be at least 1.0 goods basket.

If it less than that the sanity check fails and a sanity report is generated stating such.

To retest this set the BX.goods.pop cell to something really small and try to lock the scenario.