Fixed Bug #{3641}
Done.

Exports from the goods, black and population sectors are now displayed. They
are computed from other cells in the SAM and are, hence, outputs. These values
are then sent to the CGE when the scenario is locked.

Bring up Athena and inspect the SAM tab under the Econ tab and verify that exports
for those three sectors are visible.

Additionally you should be able to change the values in the X.i.world sectors for
i = goods, black and pop and see the exports change.